feat(config): debug mode (#65)
Introduces `--debug` flag (and `cfg.DebugMode`) to enable debugging features for development purposes.

`dev` overlay has `DEBUG_MODE` environment variable injected to enable it for local development.

With this change Gin framework runs in `production` mode, rather than `debug` (which is the default and is shown as `WARNING` in the logs). 

```
[GIN-debug] [WARNING] Creating an Engine instance with the Logger and Recovery middleware already attached.

[GIN-debug] [WARNING] Running in "debug" mode. Switch to "release" mode in production. 
- using env:    export GIN_MODE=release
- using code:    gin.SetMode(gin.ReleaseMode)
```

**Bonus**: removed duplicated "getEnvOrDefault" and swapped to k8s utils.


Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>