Migrate kuadrant to 1.3.0-alpha2 using the upstream chart (#60)
- Remove the OLM installation and use the newly cut Helm chart
- Removed a grafana datasource since the installer doesnt install
the Grafana operator anymore. (if we add it back it should be a
seperate PR). Perses is more likely to replace Grafana.
- Consolidate `deployment/scripts/teardown.sh` into
deployment/scripts/install.sh --teardown to make it easier to
maintain.
- There are some legacy uninstall bits from the OLM installation
in the teardown we can clean up later but leaving them in for now
makes sense to get everyone's clusters migrated.

Signed-off-by: Brent Salisbury <bsalisbu@redhat.com>