fix(model-listing): promotes internal address when no public one (#101)
available

When Gateway is attached to internal loadbalancer service, e.g.

```
loadBalancer:
    ingress:
    - ip: 10.0.214.36
      ipMode: VIP
```

LLMInferenceService address resolution does not set `.status.url` field,
as it is not considered public URL. The address is available in
`.status.addresses`.

With this change model listing endpoint will promote internal address in
the `/v1/models` response.

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>