docs: add MaaS manual install (#197)
* docs: add MaaS manual install

This adds the following new page to the _User Guide_ section of the documentation:
* maas-setup.md, which guides users to set up MaaS assuming ODH project or RHOAI has already been setup in the cluster.

Also, the `prerequisites.md` is slightly modified to work also as the setup overview. The navigation is updated accordingly.

The Kustomize OpenShift overlay had to be adjusted to provide setup with a minimal number of shell commands.

The `deployment/README.md` file is removed as cleanup, since its content is well covered inside the `docs` directory.

* Fix feedback: coderabbit - typo

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>