build(operator): add overlay for ODH (#314)
## Description

This adds a new `odh` Kustomize overlay. This is going to be the overlay
that ODH operator will use for MaaS setup. The `openshift` overlay is
adapted to be based on `odh` overlay, because the `openshift` one covers
more (or said in another way, which could be more meaningful: the new
odh overlay is covering less).

Despite the changes in the `openshift` overlay, the generated manifest
with `kustomize build deployment/overlays/openshift` is identical to the
one generated without these changes.

When comparing the `odh` overlay with the `openshift` one:

* The `odh` overlay won't install the Gateway, because end-users are
expected to create it prior to installing ODH.
* The `odh` overlay won't install RateLimitPolicy nor
TokenRateLimitPolicy resources, because there may be no good defaults
generally suitable. Users should bring their own policies as post-setup.
* The `odh` overlay won't install observability related resources,
because no observability features are planned for the initial
ODH-integrated release.

Essentially, the `openshift` overlay is becoming a developer overlay to
setup MaaS in a more complete way for testing and development, while the
`odh` overlay becomes the production setup.

## How Has This Been Tested?

* Run `kustomize build deployment/overlays/openshift` to ensure that it
works.
* Run `kustomize build deployment/overlays/odh` to ensure that it works.

## Merge criteria:

- [x] The commits are squashed in a cohesive manner and have meaningful
messages.
- [x] Testing instructions have been added in the PR body (for PRs
involving changes that are not immediately obvious).
- [x] The developer has manually tested the changes and verified that
the changes work


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **New Features**
  * Added new deployment overlay option for ODH environments.
* Introduced structured policy organization with separate authentication
and usage policy configurations.

* **Chores**
* Reorganized deployment policy configurations for improved modularity.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>

<!-- end of auto-generated comment: release notes by coderabbit.ai -->